refactor: split protocol contracts
